+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/services/pin_code_service.dart
@@ -1,8 +1,4 @@
 abstract class PinCodeService {
-  /// Reads from device internal storage and
-  /// returns the stored pin code to the bloc
-  Future<bool> isPinCodeInSecureStorage();
-
   /// Receives the encrypted pinCode from user input.
   /// Returns the value verified by the server
   Future<dynamic> verifyPinCode(String pinCode);
@@ -11,11 +7,13 @@ abstract class PinCodeService {
   /// the length should be less than 10 digits
   Future<int> getPinLength();
 
-  /// Encrypts the string and stores it in the device secure storage. It has a
-  /// add a default implementation to return what is the input. Returns the
-  /// encrypted pinCode from the device secure storage.
+  /// Encrypts the provided pin code and returns the encrypted string. It has a
+  /// default implementation to return what is the input.
   Future<String> encryptPinCode(String pinCode) => Future.value(pinCode);
 
   /// Returns the pin code or null if it is not saved in the device memory yet.
   Future<String?> getPinCode();
+
+  /// Stores the encrypted pin code in the device secure storage.
+  Future<bool> savePinCodeInSecureStorage(String pinCode);
 }