+++ packages/widget_toolkit_pin/test/lib_pin_code/blocs/pin_code_test.dart
@@ -30,7 +30,6 @@ void main() {
     List<BiometricsAuthType>? biometrics,
     String pinCode = '',
     bool isPinVerified = true,
-    bool isPinCodeInSecureStorage = true,
     bool enableBiometrics = true,
     BiometricsMessage biometricsMessage = BiometricsMessage.enabled,
     ErrorModel? biometricsError,
@@ -62,9 +61,6 @@ void main() {
     when(pinCodeService.verifyPinCode(pinCode))
         .thenAnswer((_) => Future.value(isPinVerified));
 
-    when(pinCodeService.isPinCodeInSecureStorage())
-        .thenAnswer((_) => Future.value(isPinCodeInSecureStorage));
-
     when(pinCodeService.encryptPinCode(pinCode))
         .thenAnswer((_) => Future.value(pinCode));
 
@@ -77,6 +73,7 @@ void main() {
         biometricAuthenticationService: biometricAuthenticationService,
         localizedReason: '',
         pinCodeService: pinCodeService,
+        autoPromptBiometric: false,
       );
 
   group('test pin_code_bloc_dart state digitsCount', () {
@@ -92,21 +89,23 @@ void main() {
         state: (bloc) => bloc.states.digitsCount,
         expect: [6]);
 
-    rxBlocTest<PinCodeBloc, int>(
+    rxBlocFakeAsyncTest<PinCodeBloc, int>(
         'test pin_code_bloc_dart state digitsCount with addDigit and deleteDigit event',
-        build: () async {
+        build: () {
           defineWhen();
           return pinCodeBloc();
         },
-        act: (bloc) async {
+        act: (bloc, fakeAsync) async {
           bloc.events.addDigit('1');
+          fakeAsync.elapse(const Duration(milliseconds: 1));
           bloc.events.deleteDigit();
           bloc.events.addDigit('2');
           bloc.events.addDigit('1');
+          fakeAsync.elapse(const Duration(milliseconds: 1));
           bloc.events.deleteDigit();
         },
         state: (bloc) => bloc.states.digitsCount,
-        expect: [0, 0, 2]);
+        expect: [0, 1, 0, 1, 2, 1]);
   });
 
   group('test pin_code_bloc_dart state authenticated', () {
@@ -132,7 +131,6 @@ void main() {
         build: () async {
           defineWhen(
             isDeviceSupported: true,
-            isPinCodeInSecureStorage: true,
             isPinVerified: true,
             canCheckBiometrics: true,
             pinCode: Stubs.pinCode,
@@ -154,7 +152,6 @@ void main() {
             pinCode: Stubs.pinCode,
             isPinVerified: true,
             enableBiometrics: true,
-            isPinCodeInSecureStorage: true,
           );
 
           return pinCodeBloc();
@@ -171,7 +168,6 @@ void main() {
           defineWhen(
             pinCode: Stubs.pinCode2,
             isPinVerified: false,
-            isPinCodeInSecureStorage: false,
             areBiometricsEnabled: false,
           );
 
@@ -187,7 +183,6 @@ void main() {
           defineWhen(
             pinCode: Stubs.pinCode2,
             isPinVerified: false,
-            isPinCodeInSecureStorage: false,
             areBiometricsEnabled: false,
             biometricsError: Stubs.error,
           );