+++ packages/widget_toolkit_pin/CHANGELOG.md
@@ -1,3 +1,13 @@
+## [0.3.0]
+Fixes and improvements:
+* Fixed a bug where biometrics authentication would not work if encryption was used
+* Added biometrics authentication to the example application
+* Added `autoPromptBiometric` parameter to `PinCodeKeyboard` to automatically authenticate with biometrics if available
+
+Breaking changes:
+* Added `savePinCodeInSecureStorage` to `PinCodeService` for saving the pin code in secure storage
+* Removed `isPinCodeInSecureStorage` method from `PinCodeService`. Now this flag is set automatically when the pin code is saved in secure storage and accessed through the `getPinCode` method
+
 ## [0.2.2]
 * Fixed a bug where a newly input pin could get deleted if the user starts typing again immediately after an error shake animation starts